real/requirements.txt
torch>=2.1.1
transformers>=4.35.2
datasets>=2.15.0
accelerate>=0.25.0
wandb>=0.16.0
numpy>=1.26.2
pandas>=2.1.3
scikit-learn>=1.3.2
tensorboard>=2.15.1
pyyaml>=6.0.1
tqdm>=4.66.1 